fix: resync UI models and services with current API DTOs, routes and enums

+++ UI/src/app/modules/pages/profile/profile.component.ts
@@ -1,14 +1,12 @@
-import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { NgxGalleryAnimation, NgxGalleryImage, NgxGalleryOptions } from '@kolkov/ngx-gallery';
 import { ToastrService } from 'ngx-toastr';
-import { take } from 'rxjs/operators';
-import { Exercise } from 'src/app/core/models/Exercise';
+import { forkJoin, of } from 'rxjs';
+import { switchMap } from 'rxjs/operators';
 import { Photo } from 'src/app/core/models/Photo';
-import { User, UserAuth } from 'src/app/core/models/User';
-import { Workout } from 'src/app/core/models/Workout';
-import { AccountService } from 'src/app/core/services/account.service';
+import { Measurements, User, UserUpdate } from 'src/app/core/models/User';
+import { Workout, WorkoutExercise } from 'src/app/core/models/Workout';
 import { UserService } from 'src/app/core/services/user.service';
 import { WorkoutService } from 'src/app/core/services/workout.service';
 
@@ -18,62 +16,77 @@ import { WorkoutService } from 'src/app/core/services/workout.service';
   styleUrls: ['./profile.component.css']
 })
 export class ProfileComponent implements OnInit {
-  user: User | undefined;
-  currentUser: UserAuth | null = null;
+  user: User | null = null;
   currentUserAcc: User | null = null;
+  measurements: Measurements = emptyMeasurements();
   workouts: Workout[] = [];
-  exercises: Exercise[] = [];
-  lastWorkout: Workout;
-  sinceLastWorkout: string;
+  exerciseNames = new Map<number, string>();
+  lastWorkout: Workout | null = null;
+  sinceLastWorkout = '';
   followers: User[] = [];
   followedUsers: User[] = [];
   galleryOptions: NgxGalleryOptions[] = [];
   galleryImages: NgxGalleryImage[] = [];
-  isCurrentUser: boolean;
-  isEditMode: boolean;
-  isFollowed: boolean;
-  fileName = "";
+  isCurrentUser = false;
+  isEditMode = false;
+  isFollowed = false;
   currentPhoto: Photo | null = null;
 
-  constructor(private userService: UserService, private accountService: AccountService,
-    private route: ActivatedRoute, private toastr: ToastrService, private workoutService: WorkoutService,
-    private http: HttpClient) {
-    this.accountService.currentUser$.pipe(take(1)).subscribe({
-      next: currentUser => this.currentUser = currentUser
-    });
-  }
+  constructor(private userService: UserService, private route: ActivatedRoute,
+    private toastr: ToastrService, private workoutService: WorkoutService) { }
 
   ngOnInit(): void {
     this.loadProfile();
-    this.loadExercises();
   }
 
   loadProfile() {
     const username = this.route.snapshot.paramMap.get('username');
-    if (!username) return;
+    if (!username)
+      return;
 
-    this.userService.getUser(username).subscribe({
-      next: user => {
+    forkJoin({
+      user: this.userService.getUser(username),
+      currentUserAcc: this.userService.getCurrentUser(),
+      exercises: this.workoutService.getExercises()
+    }).pipe(
+      switchMap(({ user, currentUserAcc, exercises }) => {
         this.user = user;
-        this.galleryImages = this.loadGallery();
-        this.loadWorkouts();
-        if (this.user.userName == this.currentUser.username)
-        {
-          this.isCurrentUser = true;
-        }
-        else
-        {
-          this.userService.getUser(this.currentUser.username).subscribe({
-            next: userAcc => this.currentUserAcc = userAcc
-          });
-        }
-        this.loadFollowedUsers();
-        this.loadFollowers();
+        this.currentUserAcc = currentUserAcc;
+        this.isCurrentUser = user.id === currentUserAcc.id;
+        this.isFollowed = user.followers.some(f => f.userId === currentUserAcc.id);
+        this.exerciseNames = new Map(exercises.map(e => [e.id, e.name ?? 'Unknown exercise']));
+        this.measurements = user.measurements ?? emptyMeasurements();
+        this.galleryImages = this.loadGallery(user);
+
+        const followerIds = user.followers.map(f => f.userId);
+        const followedIds = user.followedUsers.map(f => f.followedUserId);
+
+        return forkJoin({
+          followers: followerIds.length ? this.userService.getUsersByIds(followerIds) : of<User[]>([]),
+          followedUsers: followedIds.length ? this.userService.getUsersByIds(followedIds) : of<User[]>([]),
+          workouts: this.workoutService.getUserWorkouts(user.id)
+        });
+      })
+    ).subscribe({
+      next: ({ followers, followedUsers, workouts }) => {
+        this.followers = followers;
+        this.followedUsers = followedUsers;
+        this.setWorkouts(workouts);
       }
     });
   }
 
-  loadGallery() {
+  private setWorkouts(workouts: Workout[]) {
+    this.workouts = workouts;
+    if (!workouts.length)
+      return;
+
+    this.lastWorkout = workouts[0];
+    const dateDifference = new Date().getTime() - new Date(this.lastWorkout.date).getTime();
+    this.sinceLastWorkout = Math.round(dateDifference / (1000 * 3600 * 24)) + ' days ago';
+  }
+
+  loadGallery(user: User): NgxGalleryImage[] {
     this.galleryOptions = [
       {
         width: '500px',
@@ -82,22 +95,22 @@ export class ProfileComponent implements OnInit {
         imageAnimation: NgxGalleryAnimation.Slide,
         preview: false
       }
-    ]
+    ];
 
-    const imageUrls = [];
-    for (const photo of this.user.photos) {
-      imageUrls.push({
-        small: photo.url,
-        medium: photo.url,
-        big: photo.url
-      });
-    }
+    const imageUrls = user.photos.map(photo => ({
+      small: photo.url ?? undefined,
+      medium: photo.url ?? undefined,
+      big: photo.url ?? undefined
+    }));
 
-    if (this.user.photos.length > 0) this.currentPhoto = this.user.photos[0];
+    if (user.photos.length > 0) this.currentPhoto = user.photos[0];
     return imageUrls;
   }
 
-  onChange(data: any): void {
+  onChange(data: { index: number }): void {
+    if (!this.user)
+      return;
+
     this.currentPhoto = this.user.photos[data.index];
   }
 
@@ -106,88 +119,94 @@ export class ProfileComponent implements OnInit {
   }
 
   saveChanges() {
-    this.userService.updateUser(this.user).subscribe({
-      next: _ => {
+    const user = this.user;
+    if (!user)
+      return;
+
+    const update: UserUpdate = {
+      id: user.id,
+      dateOfBirth: user.dateOfBirth,
+      dateOfWorkoutStart: user.dateOfWorkoutStart,
+      name: user.name,
+      surname: user.surname,
+      gender: user.gender,
+      city: user.city,
+      country: user.country,
+      description: user.description,
+      measurements: this.measurements,
+      photos: user.photos
+    };
+
+    this.userService.updateUser(update).subscribe({
+      next: updated => {
+        this.user = updated;
+        this.measurements = updated.measurements ?? emptyMeasurements();
         this.changeEditMode();
         this.toastr.success('Profile has been updated');
       }
     });
   }
 
-  loadFollowedUsers() {
-    var followedUsersIds = [];
-    this.user.followedUsers.forEach(x => {
-      followedUsersIds.push(x.followedUserId);
-    });
-    this.userService.getFollowerUsers(followedUsersIds).subscribe({
-      next: users => {
-        this.followedUsers = users;
-      }
-    })
+  getExerciseName(exerciseId: number) {
+    return this.exerciseNames.get(exerciseId) ?? 'Unknown exercise';
   }
 
-  loadFollowers() {
-    var followersIds = [];
-    this.user.followers.forEach(x => {
-      followersIds.push(x.userId);
-    });
-    this.userService.getFollowerUsers(followersIds).subscribe({
-      next: users => {
-        this.followers = users;
-        if (followersIds.includes(this.currentUserAcc.id)) this.isFollowed = true;
-      }
-    })
+  formatReps(exercise: WorkoutExercise) {
+    const reps = exercise.sets.map(s => s.reps);
+    if (!reps.length) return '-';
+
+    return reps.every(r => r === reps[0])
+      ? `${reps.length}x${reps[0]}`
+      : reps.join(' / ');
   }
 
-  loadWorkouts() {
-    this.workoutService.getUserWorkouts(this.user).subscribe({
-      next: workouts => {
-        this.workouts = workouts;
-        if (this.workouts.length != 0) {
-          this.lastWorkout = this.workouts[0];
-          var dateDifference = new Date().getTime() - new Date(this.lastWorkout.date).getTime();
-          this.sinceLastWorkout = Math.round(dateDifference / (1000 * 3600 * 24)) + ' days ago';
-        }
-      }
+  formatWeight(exercise: WorkoutExercise) {
+    const weights = exercise.sets.map(s => s.weight ?? 0);
+    if (!weights.length)
+      return '-';
+
+    return weights.every(w => w === weights[0])
+      ? `${weights[0]} kg`
+      : `${weights.join(' / ')} kg`;
+  }
+
+  toggleFollow() {
+    if (!this.user || !this.currentUserAcc)
+      return;
+
+    const request = this.isFollowed
+      ? this.userService.unfollowUser(this.currentUserAcc.id, this.user.id)
+      : this.userService.followUser(this.currentUserAcc.id, this.user.id);
+
+    request.subscribe({
+      next: _ => location.reload()
     });
   }
 
-  loadExercises() {
-    this.workoutService.getExercises().subscribe({
-      next: exercises => this.exercises = exercises
-    })
-  }
+  onFileSelected(event: Event) {
+    if (!this.user)
+      return;
 
-  getExerciseName(id: number) {
-    return this.exercises.find(x => x.id == id).name;
-  }
+    const input = event.target as HTMLInputElement;
+    const file = input.files?.[0];
+    if (!file)
+      return;
 
-  followUser() {
-    return this.userService.followUser(this.currentUserAcc.id, this.user.id).subscribe({
+    const formData = new FormData();
+    formData.append('file', file);
+    this.userService.addPhoto(formData, this.user.id).subscribe({
       next: _ => {
         location.reload();
+        this.toastr.success('Photo has been added');
       }
     });
-  }
-
-  onFileSelected(event) {
-    const file:File = event.target.files[0];
-
-    if (file) {
-        this.fileName = file.name;
-        const formData = new FormData();
-        formData.append('file', file);
-        return this.userService.addPhoto(formData, this.user.id).subscribe({
-          next: _ => {
-            location.reload();
-            this.toastr.success('Photo has been added');
-          }
-        });
-    }
   }
 
   setMainPhoto() {
-    return this.userService.setMainPhoto(this.currentPhoto.id, this.user.id).subscribe({
+    if (!this.currentPhoto || !this.user)
+      return;
+
+    this.userService.setMainPhoto(this.currentPhoto.id, this.user.id).subscribe({
       next: _ => {
         location.reload();
         this.toastr.success('Main photo has been changed');
@@ -196,11 +215,27 @@ export class ProfileComponent implements OnInit {
   }
 
   deletePhoto() {
-    return this.userService.deletePhoto(this.currentPhoto.id, this.user.id).subscribe({
+    if (!this.currentPhoto || !this.user)
+      return;
+
+    this.userService.deletePhoto(this.currentPhoto.id, this.user.id).subscribe({
       next: _ => {
         location.reload();
         this.toastr.success('Photo has been deleted');
       }
     });
   }
+}
+
+function emptyMeasurements(): Measurements {
+  return {
+    height: null,
+    weight: null,
+    chest: null,
+    shoulders: null,
+    arms: null,
+    waist: null,
+    hips: null,
+    thights: null
+  };
 }
